Trigger auto load-more on sentinel visibility instead of timer loop
The auto-load mode previously fired the load action every 300 ms in a
tight task loop while the sentinel was rendered, which paginated through
the entire feed regardless of scroll position. Drive it from
onScrollVisibilityChange so it triggers exactly when the user scrolls
the sentinel into view, and re-triggers naturally when it scrolls back
in after the next batch.

Also extracts LoadPreviousArticlesButton into its own file.
