+++ SakuraRSS/Views/Shared/Articles/ArticlesView.swift
@@ -215,51 +215,3 @@ struct ArticlesView: View {
         return displayStyle
     }
 }
-
-struct LoadPreviousArticlesButton: View {
-
-    let action: () -> Void
-
-    @AppStorage("Articles.AutoLoadWhileScrolling") private var autoLoadWhileScrolling: Bool = false
-
-    var body: some View {
-        Group {
-            if autoLoadWhileScrolling {
-                HStack(spacing: 8) {
-                    ProgressView()
-                    Text(String(localized: "LoadPrevious.Loading", table: "Articles"))
-                        .font(.subheadline)
-                        .foregroundStyle(.secondary)
-                }
-                .frame(maxWidth: .infinity)
-                .padding(.vertical, 12)
-                .task {
-                    while !Task.isCancelled {
-                        await MainActor.run {
-                            withAnimation(.smooth.speed(2.0)) {
-                                action()
-                            }
-                        }
-                        try? await Task.sleep(for: .milliseconds(300))
-                    }
-                }
-            } else {
-                Button {
-                    withAnimation(.smooth.speed(2.0)) {
-                        action()
-                    }
-                } label: {
-                    HStack {
-                        Image(systemName: "clock.arrow.circlepath")
-                        Text(String(localized: "LoadPrevious", table: "Articles"))
-                    }
-                    .font(.subheadline)
-                    .frame(maxWidth: .infinity)
-                    .padding(.vertical, 12)
-                }
-                .buttonStyle(.bordered)
-                .tint(.secondary)
-            }
-        }
-    }
-}
